Adding goap planner (#7)
* Added basic debugging support | Fixed the bug where astar was not able to find correct path

* Removed logs because they were taking too much time and caused drop in framerate

* multithreaded pop of heap that matches the single threaded implementation | has bugs

* Adding changes for goap planner

+++ AstarPathGame/Assets/MainGame/EditModeTests/Helpers/DirectedEdge.cs
@@ -4,7 +4,7 @@ namespace TestHelpers
 {
     public class DirectedEdge : IWeightedEdge<LocationNode>
     {
-        public double Weight { get; set; }
+        public float Weight { get; set; }
 
         public LocationNode OriginNode { get; set; }
         public LocationNode DestinationNode { get; set; }
@@ -12,7 +12,7 @@ namespace TestHelpers
         public DirectedEdge()
         { }
 
-        public DirectedEdge(LocationNode originNode, LocationNode destinationNode, double cost)
+        public DirectedEdge(LocationNode originNode, LocationNode destinationNode, float cost)
         {
             OriginNode = originNode;
             DestinationNode = destinationNode;
